refactor: remove masonry in favor of flex
