--- src/lib/components/gallery/grid.module.css
@@ -1,92 +1,0 @@
-.container {
-	display: flex;
-	flex-flow: column nowrap;
-	gap: var(--space-m);
-}
-
-.control-panel > label {
-	width: min-content;
-	grid-auto-flow: column;
-}
-
-/* flex gallery view */
-
-.gallery {
-	display: flex;
-	flex-flow: row wrap;
-	gap: var(--space-m);
-	width: 100%;
-}
-
-.gallery::after {
-	/* do not let stretch the last item */
-	content: "";
-	flex-grow: 10;
-}
-
-.gallery-item {
-	flex-grow: 1;
-	height: var(--grid-item-size, 250px);
-	width: auto;
-	position: relative;
-}
-
-/* masonry view */
-
-.gallery:global(.masonry) {
-	display: grid;
-	grid-template-columns: repeat(auto-fill, minmax(var(--masonry-item-size, 50px), 1fr));
-	grid-template-rows: repeat(auto-fit, var(--cell-size, 50px));
-	grid-auto-flow: dense;
-	place-items: stretch;
-	gap: var(--space-m);
-	width: 100%;
-}
-
-.gallery:global(.masonry) > .gallery-item {
-	grid-row-end: var(--masonry-item-row);
-	grid-column-end: var(--masonry-item-column);
-	height: 100%;
-}
-
-@media (max-aspect-ratio: 1 / 1) {
-  .gallery-item {
-    height: 30vh;
-  }
-}
-
-@media (max-height: 480px) {
-  .gallery-item {
-		--grid-item-size: 80vh;
-  }
-}
-
-@media (max-aspect-ratio: 1/1) and (max-width: 480px) {
-  .gallery {
-    flex-direction: row;
-  }
-
-  .gallery-item {
-    height: auto;
-    width: 100%;
-  }
-
-  .gallery-item img {
-    width: 100%;
-    max-height: 75vh;
-    min-width: 0;
-  }
-}
-
-/* masonry media */
-@media screen and (max-width: 599px) {
-	.gallery:global(.masonry) {
-		grid-template-columns: 1fr;
-		grid-template-rows: unset;
-		grid-auto-columns: auto;
-	}
-
-	.gallery:global(.masonry) > .gallery-item {
-		grid-column-end: unset;
-	}
-}
