+++ src/lib/components/gallery/index.ts
@@ -1,3 +1,2 @@
 export { default as Gallery } from "./Gallery.svelte";
-export { default as GalleryGrid } from "./Grid.svelte";
 export type { GalleryItem } from "./Gallery.types";
